Obtaining the metrics.

Key metrics

The key metrics for MHWs or MCSs as defined by Hobday et al. are contained in a const metrics which is a dictionary of key-value pairs.

These metrics are calculated from the Events which is passed in as an argument to meanmetrics and annualmetrics.

Obtaining the trend from trend requires the output from annualmetrics.

MarineHeatwaves.meanmetricsm — Function

Return the mean of the metrics in each pixel.

MarineHeatwaves.annualmetricsm — Function

Compute the metrics for each year in the desired period.

MarineHeatwaves.mmetricsm — Function

Return the computed metrics - means, annuals and linear regression outputs as MHCMetrics.

MarineHeatwaves.mymetric — Function

Return the Eventsfull of a single pixel as a matrix

Return all the values of the events of a given metric as a vector.

mymetric(eve, "means")::Vector{T}

Return the Events be used as a table (or dataframe).

mymetric(evt, indices, startends)::Matrix

Return the climatology mean and threshold as input array.

ma = mymetric(mm, arg, indices)

MarineHeatwaves.linreg — Function

a,b,r2,sigma_a,sigma_b,sigma_e = linreg(x,y)

Linear regression

y ≈ a + b * x

Multi-dimensional linear regression y ≈ a + X b

a is a scalar and b a vector

y = a + X b + ϵ

Σ ϵ² = (y- (a + X b))^2

y = X beta + ϵ

beta = inv(X*X') * X' * y