Events

Types

MarineHeatwaves.MEvents — Type

MEvents{TE}

Abstract type for CS and HW Events.

MarineHeatwaves.EventCS — Type

EventCS{TE}: event wrapper for cold spell.

MarineHeatwaves.EventHW — Type

EventHW{TE}: event wrapper for heatwaves.

MarineHeatwaves.Events — Type

Events{TE}(means, mini/max, onset, decline, duration, sums, categorys, anomaly, threshanomaly, dt::MEvents)

Functions and methods

MarineHeatwaves.mylabel — Function

Return the start and stop positions and pixel columns at which MExtremes were detected.

mylabel(me::MExtreme, minimum_duration, maximum_gap)

MarineHeatwaves.anomsam — Function

Compute the anomaly of the events in each pixel and return the Event.

MarineHeatwaves.meventsm — Function

meventsm(ms, min_duration, max_gap)

Return the Events and the labels (starts and end positions) of the event.

MarineHeatwaves.mevents — Function

mevents(anomaly) -> Return tuple of means, sums, maximums and durations

MarineHeatwaves.decline — Function

Return the rate of decline for each event.

MarineHeatwaves.onset — Function

Return the rate of onset for each event.